% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{data_for_annotation}
\alias{data_for_annotation}
\title{Prepare the query and reference data for annotation}
\usage{
data_for_annotation(obj, panel.data, min.express = 0, verbose = TRUE)
}
\arguments{
\item{obj}{Seurat object for annotation}

\item{panel.data}{Reference panel data}

\item{min.express}{Filter for genes expressed in at least this
proportions of samples in the obj. Expressed between 0 to 1.}

\item{verbose}{}
}
\value{
A list containing the filtered query and filtered reference panel.
}
\description{
Extracts the data assay from a Seurat object and
optionally filters out genes that are expressed in a minimum number of samples.
It subsets the filtered data and the reference panel with
genes present in both datasets.
}
