% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.demo}
\alias{pbmc.demo}
\title{Demo dataset prepared from the PBMC3k}
\format{
A Seurat object with 13,714 genes and 2,635 cells. The meta data slot contains:
\describe{
  \item{nCount_RNA}{Number of UMIs for each cell}
  \item{nFeature_RNA}{Number of genes detected for each cell}
  \item{percent.mt}{Percentage of reads mapping to mitochondrial genes per cell}
  \item{seurat_annotations}{Cell type annotation provided by authors of Seurat }
}
}
\source{
pbmc3k data from the SeuratData package
}
\usage{
pbmc.demo
}
\description{
The famous PBMC3k dataset used in the main Seurat tutorial
after quality control filtering and after standard preprocessing:
log normalization, highly variable genes, scaling, PCA and UMAP.
The dimensionality of the data was set to 10 from Elbow plot inspection.
}
\examples{
pbmc.demo@meta.data \%>\% tabyl(seurat_annotations)

DimPlot(pbmc.demo,
        reduction = "umap",
        group.by  = "seurat_annotations",
        label = TRUE) +
  NoLegend()

}
\keyword{datasets}
