% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{RCAannotate}
\alias{RCAannotate}
\title{Run RCAv2 annotation}
\usage{
RCAannotate(obj, panel.name, ...)
}
\arguments{
\item{obj}{Seurat object for annotation}

\item{panel.name}{Name of the reference panel.}

\item{...}{Optional parameters. Currently only `min.express` and `power` is recognized.}
}
\value{
A data frame with the predicted cell types
}
\description{
Run RCAv2 annotation
}
